Removed monthly logging
Now just keeps track of each project in a separate file. For now we only
have the default cisco project, but others should be added later.

Overal the error handling in the parser should be vastly improved.

Also we cant use it now as we have to initialize the project before
logging to it.