Add printouts of flex and log when logging
This is the main feature done! If we just don't want to do anything
except log time every weekday in a slightly wonky manner it works like a
charm! Just don't try anything else there!

The code is not great. Especially the time module is an unnecessary
dumpster fire just because I thought it would be fun to not use chrono
for everything...